cleanup api stuff, general file extension fixes
