+++ package.json
@@ -9,6 +9,7 @@
     "@types/jest": "^25.2.3",
     "@types/react": "^16.9.35",
     "@types/react-native": "^0.62.10",
+    "@types/react-native-dotenv": "^0.2.0",
     "@types/react-redux": "^7.1.9",
     "FormData": "^0.10.1",
     "awesome-phonenumber": "^2.33.0",
@@ -29,8 +30,10 @@
     "react-addons-test-utils": "^15.6.2",
     "react-dom": "^16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
+    "react-native-dotenv": "^0.2.0",
     "react-native-dropdownalert": "^4.2.1",
     "react-native-elements": "^2.0.0",
+    "react-native-fs": "^2.16.6",
     "react-native-gesture-handler": "^1.6.1",
     "react-native-reanimated": "^1.8.0",
     "react-native-safe-area-context": "^1.0.2",
