+++ src/views/index.ts
